requirements.txt
gunicorn==20.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dominate==2.7.0
Flask==2.1.3
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.4.1
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==1.0.1
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.3.19
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
